uclibc-compat: split .so and .a to fix process-wide symbol interposition
libuclibc-compat.so was exporting stat/mmap which overrode musl's
implementations for ALL code in the process, including musl's own
internal calls.  This caused segfaults because musl expects 64-bit
off_t but the shim's mmap takes uint32_t.

Split into two libraries:
- libuclibc-compat.so: only symbols MISSING from musl (__ctype_b,
  __assert, __fgetc_unlocked, memcpy_s, etc).  Safe as shared lib.
- libuclibc-compat-static.a: stat/mmap wrappers with uclibc-compatible
  ABI.  Must be statically linked into the executable so overrides
  only affect vendor .so lookups through the executable's symbol scope.

BUNDLE_SDK now builds and includes both in the toolchain tarball.

Fixes: segfault in majestic on hi3516cv100 at hisi_free_mem_ex (mmap)
Ref: #1992
