+++ general/package/uclibc-compat/src/Makefile
@@ -1,12 +1,16 @@
-LIB := libuclibc-compat
-
 CFLAGS ?= -O2
 CFLAGS += -Wall -fPIC
 
-all: $(LIB).so
+all: libuclibc-compat.so libuclibc-compat-static.a
 
-$(LIB).so: uclibc-compat.c
+libuclibc-compat.so: uclibc-compat.c
 	$(CC) $(CFLAGS) -shared -o $@ $^
 
+libuclibc-compat-static.a: uclibc-compat-static.o
+	$(AR) rcs $@ $^
+
+uclibc-compat-static.o: uclibc-compat-static.c
+	$(CC) $(CFLAGS) -c -o $@ $<
+
 clean:
-	rm -f $(LIB).so
+	rm -f *.o *.so *.a
